update the scripts and seperate tests from scripts
